Removed VirtualDom and HTML parser
They worked well enough to be somewhat stable, although not tested against edge cases. Removing because extensive testing and benchmarking shows that a VDOM will only allow for very marginal gains in performance and only under specific conditions in modern browsers. Not enough to warrant complicating the foundation for future UI development in Accede. Should it be needed in the future, the code is here and working in a mostly usable state.